Unified naming across the RealmBrowser component
