+++ src/ui/RealmBrowser/Table/types/StringCell.tsx
