+++ src/ui/RealmBrowser/Table/types/StringCellContainer.tsx
