--- src/ui/realm-browser/index.ts
@@ -1,30 +1,0 @@
-import { IClassFocus } from './focus';
-
-// TODO: Remove this interface once the Realm.ObjectSchemaProperty
-// has a name parameter in its type definition.
-export interface IPropertyWithName extends Realm.ObjectSchemaProperty {
-  name: string | null;
-  readOnly: boolean;
-}
-
-export interface ISelectObjectState {
-  focus: IClassFocus;
-  property: IPropertyWithName;
-  contentToUpdate: Realm.Object | Realm.List<any>;
-}
-
-export enum EditMode {
-  Disabled = 'disabled',
-  InputBlur = 'input-blur',
-  KeyPress = 'key-press',
-}
-
-export type CreateObjectHandler = (className: string, values: {}) => void;
-export type EditModeChangeHandler = (editMode: EditMode) => void;
-
-export interface IConfirmModal {
-  title: string;
-  description: string;
-  yes: () => void;
-  no: () => void;
-}
